\name{HW5MCMC-package}
\alias{HW5MCMC-package}
\alias{HW5MCMC}
\docType{package}
\title{
\packageTitle{HW5MCMC}
}
\description{
\packageDescription{HW5MCMC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW5MCMC}
\packageIndices{HW5MCMC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HW5MCMC}

Maintainer: \packageMaintainer{HW5MCMC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
